tests/wpt/tests/dom/nodes/Element-nextElementSibling-svg.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     version="1.1"
     width="100%" height="100%" viewBox="0 0 400 400">
<title>nextElementSibling</title>
<h:script src="/resources/testharness.js"/>
<h:script src="/resources/testharnessreport.js"/>

<text x="200" y="40" font-size="25" fill="black" text-anchor="middle">Test of nextElementSibling</text>
<text id="parentEl" x="200" y="70" font-size="20" fill="black" text-anchor="middle">The result of <tspan id="first_element_child">this test</tspan> is <tspan id="last_element_child" font-weight="bold">unknown.</tspan></text>

<h:script><![CDATA[
test(function() {
  var parentEl = document.getElementById("parentEl");
  var fec = document.getElementById("first_element_child");
  var nes = fec.nextElementSibling;
  assert_true(!!nes)
  assert_equals(nes.nodeType, 1)
  assert_equals(nes.getAttribute("id"), "last_element_child")
})
]]></h:script>
</svg>
